fix memcpy

+++ Lib/App/Sgemm.cpp
@@ -60,8 +60,8 @@ void conv1x1s1_sgemm_qpulib(Ptr<Float> bottom, Ptr<Float> top, Ptr<Float> kernel
 
             If (last_i + 1 < w * h)
                 Int left_len = w * h - last_i - 1;
-                Int all_one = 1;
-                Int all_zero = 0;
+                Float all_one = 1.0f;
+                Float all_zero = 0;
 
                 Where(index() < left_len)
                     all_zero = all_one;
